agency 관련 커밋 #67

+++ src/main/java/com/tsp/new_tsp_admin/api/model/service/agency/AdminAgencyJpaRepository.java
@@ -134,7 +134,7 @@ public class AdminAgencyJpaRepository {
      * 5. 작성일       : 2022. 08. 14.
      * </pre>
      */
-    public Integer deleteAgency(Integer idx) {
+    public Long deleteAgency(Long idx) {
         em.remove(em.find(AdminAgencyEntity.class, idx));
         em.flush();
         em.clear();
